refactor learning screen
